requirements.txt
progressbar2==3.47.0
Flask==1.1.1
Flask-Assets==2.0
Flask-Mail==0.9.1
sqlalchemy_utils==0.36.1
SQLAlchemy>=0.9
Flask-SQLAlchemy==2.4.1
jsmin==2.2.2
Flask-OpenID==1.2.5
Flask-OAuth==0.12
pytz==2019.3
bcrypt>=2.0.0
ordereddict==1.1
Flask-WTF==0.14.3
requests==2.22.0
blinker==1.4
Flask-RQ2==18.3
tweepy==3.8.0
oauth2client==4.1.3
ua-parser==0.8.0
itsdangerous==1.1.0
psycopg2==2.8.4
phonenumbers==8.11.2
git+https://github.com/hasgeek/coaster
git+https://github.com/hasgeek/baseframe
